Fix stylistic lint errors

+++ pkg/collector/metrics.go
@@ -649,7 +649,7 @@ func (c *DataMetricsCollector) Collect(ch chan<- prometheus.Metric) {
 		}
 
 		behavior := c.Cluster.BehaviorForResource(serviceType, resourceName, "")
-		overcommitFactor := float64(behavior.OvercommitFactor)
+		overcommitFactor := behavior.OvercommitFactor
 		if overcommitFactor == 0 {
 			overcommitFactor = 1
 		}
